requirements.txt
streamlit
pandas
pydeck
folium
streamlit-folium
requests
edge-tts
google-generativeai
